refactor: move safety composition to runtime
