+++ crates/cn-indexer/tests/ingestion_contracts.rs
@@ -8,9 +8,7 @@
 use std::sync::Arc;
 
 use anyhow::Result;
-use kukuri_cn_core::{
-    IndexScopeKind, MemoryIndexEntryStore, MemorySafetyArtifactStore, SafetyScanService,
-};
+use kukuri_cn_core::{IndexScopeKind, MemoryIndexEntryStore};
 use kukuri_cn_indexer::config::RelayConfig;
 use kukuri_cn_indexer::ingest::IngestPipeline;
 use kukuri_cn_indexer::participant::ScopeReplica;
@@ -21,6 +19,7 @@ use kukuri_cn_safety::{
 };
 use kukuri_cn_safety_runtime::clock::SystemScanClock;
 use kukuri_cn_safety_runtime::id::UuidEventIdGenerator;
+use kukuri_cn_safety_runtime::{MemorySafetyArtifactStore, SafetyScanService};
 use kukuri_cn_safety_runtime::{
     SafetyOrchestrator, Secp256k1ModerationEventSigner, verify_signed_event,
 };
